Move to splineutil2.h all the prototypes of non-static splineutil2.c functions
