+++ fontforge/Makefile.am
@@ -84,6 +84,7 @@ noinst_HEADERS = \
 	splinesaveafm.h \
 	splinesave.h \
 	splinestroke.h \
+	splineutil2.h \
 	fontforgeui.h sftextfieldP.h configure-fontforge.h views.h 	\
 	flaglist.h strlist.h charview_private.h cvundoes.h tables.h
