+++ fontforge/autohint.c
@@ -36,6 +36,7 @@
 #include "splinefont.h"
 #include "splinefill.h"
 #include "splinesave.h"
+#include "splineutil2.h"
 #include "views.h"
 #include "stemdb.h"
 #include <utype.h>
