+++ fontforge/cvimages.c
@@ -34,6 +34,7 @@
 #include "psread.h"
 #include "spiro.h"
 #include "splineorder2.h"
+#include "splineutil2.h"
 #include <math.h>
 #include <sys/types.h>
 #include <dirent.h>
