+++ fontforge/cvundoes.c
@@ -40,6 +40,7 @@
 #include "spiro.h"
 #include "splinefill.h"
 #include "splineorder2.h"
+#include "splineutil2.h"
 #include "views.h"
 #include <math.h>
 #include <ustring.h>
