+++ fontforge/effects.c
@@ -32,6 +32,7 @@
 #include "fontforgevw.h"
 #include "splineoverlap.h"
 #include "splinestroke.h"
+#include "splineutil2.h"
 #include <ustring.h>
 #include <gkeysym.h>
 #include <utype.h>
