+++ fontforge/fvimportbdf.c
@@ -39,6 +39,7 @@
 #include "palmfonts.h"
 #include "parsettf.h"
 #include "splinefill.h"
+#include "splineutil2.h"
 #include <gfile.h>
 #include <math.h>
 #include "utype.h"
