+++ fontforge/palmfonts.c
@@ -34,6 +34,7 @@
 #include "mem.h"
 #include "splinefill.h"
 #include "splinesaveafm.h"
+#include "splineutil2.h"
 #include <stdio.h>
 #include <math.h>
 #include "splinefont.h"
