+++ fontforge/scripting.c
@@ -68,6 +68,7 @@
 #include "splineorder2.h"
 #include "splinesaveafm.h"
 #include "splinestroke.h"
+#include "splineutil2.h"
 #include <gfile.h>
 #include <utype.h>
 #include <ustring.h>
