+++ fontforge/search.c
@@ -30,6 +30,7 @@
 #include "cvundoes.h"
 #include "fontforgevw.h"
 #include "fvfonts.h"
+#include "splineutil2.h"
 #include <math.h>
 #include <ustring.h>
 #include <utype.h>
