+++ fontforge/splineorder2.c
@@ -29,6 +29,7 @@
 
 #include "fontforge.h"
 #include "splinerefigure.h"
+#include "splineutil2.h"
 #include <math.h>
 #include <unistd.h>
 #include <time.h>
