+++ fontforge/splinerefigure.c
@@ -31,6 +31,7 @@
 #include <stdio.h>
 #include <math.h>
 #include "splinefont.h"
+#include "splineutil2.h"
 #ifdef HAVE_IEEEFP_H
 # include <ieeefp.h>		/* Solaris defines isnan in ieeefp rather than math.h */
 #endif
