+++ fontforge/splineutil.c
@@ -41,6 +41,7 @@
 #include "splinefill.h"
 #include "splineorder2.h"
 #include "splinerefigure.h"
+#include "splineutil2.h"
 #include "ustring.h"
 #include "utype.h"
 #include "views.h"		/* for FindSel structure */
