+++ fontforge/splineutil2.c
@@ -25,6 +25,9 @@
  * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+
+#include "splineutil2.h"
+
 #include "autohint.h"
 #include "cvundoes.h"
 #include "fontforge.h"
