+++ fontforge/svg.c
@@ -37,6 +37,7 @@
 #include "psread.h"
 #include "splineorder2.h"
 #include "splinesaveafm.h"
+#include "splineutil2.h"
 #include <unistd.h>
 #include <math.h>
 #include <locale.h>
