+++ fontforge/tottf.c
@@ -41,6 +41,7 @@
 #include "splineorder2.h"
 #include "splinesaveafm.h"
 #include "splinesave.h"
+#include "splineutil2.h"
 #include <math.h>
 #include <unistd.h>
 #include <time.h>
