+++ fontforge/ufo.c
@@ -41,6 +41,7 @@
 #include "fvfonts.h"
 #include "lookups.h"
 #include "splinesaveafm.h"
+#include "splineutil2.h"
 #include <unistd.h>
 #include <math.h>
 #include <time.h>
