+++ fontforgeexe/charview.c
@@ -46,6 +46,7 @@
 #include "splineorder2.h"
 #include "splineoverlap.h"
 #include "splinesaveafm.h"
+#include "splineutil2.h"
 #include <math.h>
 #include <locale.h>
 #include <ustring.h>
