+++ fontforgeexe/cvgetinfo.c
@@ -30,6 +30,7 @@
 #include "parsettf.h"
 #include "spiro.h"
 #include "splineorder2.h"
+#include "splineutil2.h"
 #include <ustring.h>
 #include <math.h>
 #include <utype.h>
