+++ fontforgeexe/cvpalettes.c
@@ -29,6 +29,7 @@
 #include "spiro.h"
 #include "splinefill.h"
 #include "splineorder2.h"
+#include "splineutil2.h"
 #include "collabclientui.h"
 
 int palettes_docked=1;
