+++ fontforgeexe/cvpointer.c
@@ -28,6 +28,7 @@
 #include "cvundoes.h"
 #include "fontforgeui.h"
 #include "spiro.h"
+#include "splineutil2.h"
 #include <utype.h>
 #include <math.h>
 #include "collabclient.h"
