+++ fontforgeexe/cvshapes.c
@@ -28,6 +28,7 @@
 #include "fontforgeui.h"
 #include "spiro.h"
 #include "splineorder2.h"
+#include "splineutil2.h"
 #include <math.h>
 
 static struct shapedescrip {
