+++ fontforgeexe/cvstroke.c
@@ -28,6 +28,7 @@
 #include "fontforgeui.h"
 #include "fvfonts.h"
 #include "splinestroke.h"
+#include "splineutil2.h"
 #include <ustring.h>
 #include <utype.h>
 #include <gkeysym.h>
