+++ fontforgeexe/fontinfo.c
@@ -40,6 +40,7 @@
 #include "sfd.h"
 #include "splineorder2.h"
 #include "splinesaveafm.h"
+#include "splineutil2.h"
 #include <ustring.h>
 #include <chardata.h>
 #include <utype.h>
