+++ fontforgeexe/problems.c
@@ -34,6 +34,7 @@
 #include "splineorder2.h"
 #include "splineoverlap.h"
 #include "splinesaveafm.h"
+#include "splineutil2.h"
 #include <gwidget.h>
 #include <ustring.h>
 #include <math.h>
