+++ fontforgeexe/tilepath.c
@@ -25,6 +25,7 @@
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 #include "fontforgeui.h"
+#include "splineutil2.h"
 #include <math.h>
 #include <gkeysym.h>
